An attempt at fixing the token stack issue.
When tokens are merged into groups it works perfectly, updating the references so that tokens are carried with ships after merging.

When tokens are removed however, TTS doesn't tell us what the GUID of the token that is left is - which is rather frustrating! So I don't know how to reattach the last remaining token to the ship without doing the usual proximity checks.